Moved classes to new packages for clarity of purpose
