+++ src/main/java/com/royware/corona/dashboard/services/chartlistmakers/ChangeInTotalCasesVersusCasesWithExponentialLineChartList.java
@@ -11,7 +11,6 @@ import org.springframework.stereotype.Component;
 
 import com.royware.corona.dashboard.interfaces.charts.IChartList;
 import com.royware.corona.dashboard.interfaces.model.CanonicalCaseDeathData;
-import com.royware.corona.dashboard.services.charts.ChartListMakerUtilities;
 
 @Component
 public class ChangeInTotalCasesVersusCasesWithExponentialLineChartList implements IChartList {
